Add Spanish translations
I've tried to make the translations as generic-Spanish as
possible. I've found a duplicate (`sponsors`) so I've removed.

I haven't translated yet `venue` as I'm looking for the proper word in
Spanish for it. In Argentina we would very likely use venue, but
that's definitely not the right word.

Last but not least, I haven't translated yet `under license` as I'd
like to know more about the context of where it is used, as it could
mean _without license_ or _using the following license_.